pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ytmanager"
version = "0.1.0"
description = "개인 YouTube 채널 업로드 영상 메타데이터를 관리하는 한국어 데스크톱 앱"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Proprietary" }
authors = [{ name = "YTmanager" }]
dependencies = [
  "PySide6>=6.7",
  "google-api-python-client>=2.140",
  "google-auth-httplib2>=0.2",
  "google-auth-oauthlib>=1.2",
  "keyring>=25",
]

[project.optional-dependencies]
dev = [
  "pyinstaller>=6.6",
]

[project.scripts]
ytmanager = "ytmanager.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
ytmanager = ["py.typed"]
